+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.2.1
+
+  - sanitize build neocore script
+  - add bootstrap create project script
+
 ## 2.2.0
 
   - update commands in readme for create a project